Passing lists to creole-wiki instead of strings
I was getting ert test failures - and found that changing the calls to
creole-wiki - passing in lists instead of strings for :css and
:javascript caused the tests to pass. Whether this means that the whole
system will come crashing down, who knows? - but it looks like the
intention in wiki-creole is to be able to pass multiple style, css and
javascript blocks in - in which case, lists certainly make a lot of
sense.